requirements.txt
python-telegram-bot[ext]==21.6
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]>=2.0.40
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.1
weasyprint==63.1
jinja2==3.1.5
google-api-python-client==2.159.0
google-auth==2.37.0
pydantic-settings==2.7.1
python-dateutil==2.9.0
